generate<vector>(int) -> generate<vector>(size_t) #15

+++ testlib/generators/vector.hpp
@@ -4,10 +4,10 @@ template<typename T>
 class DefaultGenerator<std::vector<T>> : Generator<std::vector<T>> {
 public:
 	template<typename... Args>
-	std::vector<T> generate(Random& rnd, int n, Args&&... args){
+	std::vector<T> generate(Random& rnd, size_t n, Args&&... args){
 		std::vector<T> result;
 		result.reserve(n);
-		for(int i = 0; i < n; ++i){
+		for(size_t i = 0; i < n; ++i){
 			result.push_back(rnd.next<T>(std::forward<Args>(args)...));
 		}
 		return result;
